Refactor: extract adaptors.OIDC
